Refactor children-finding methods to return nil.
Using an early return instead of a result variable gets rid of two
breaks at different nesting levels.

 I'm not actually sure if the guard against returning an empty list of
children should be preserved or if we should return nil immediately. It
depends on whether we want to allow for an item to be present multiple
times in the TOC. But if it is only the first found entry with children
will be used so the behavior is still ill-defined.

Even without the instant return in the case that an item isn't found or
has [no children][1], this does change behavior to always return nil.

[1]: https://www.youtube.com/watch?v=wRP6egIEABk